feat: Moves application settings to a separate screen
It is only available in dev version
