+++ app/frontend/src/components/Sidebar/Icons/TV.svelte
@@ -1,0 +1,1 @@
+<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" fill="none" viewBox="0 0 24 24"><path fill="var(--sidebar-icon-color)" d="M4.497 17.826h14.999c1.37 0 2.087-.693 2.087-2.086V7.046c0-1.394-.716-2.087-2.087-2.087h-15c-1.37 0-2.086.693-2.086 2.087v8.694c0 1.393.716 2.086 2.087 2.086Zm.098-1.461c-.498 0-.724-.203-.724-.716V7.136c0-.512.226-.715.724-.715h14.803c.497 0 .723.203.723.715v8.513c0 .513-.226.716-.723.716H4.595Zm3.548 4.098h7.707a.688.688 0 0 0 .685-.685.695.695 0 0 0-.685-.694H8.143a.695.695 0 0 0-.686.694c0 .376.31.685.686.685Z"/></svg>
